requirements.txt
numpy==1.26.4
chromadb==1.0.15
Flask==3.1.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
openai==1.93.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tiktoken==0.9.0
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary
Flask-Session
pdfplumber
python-docx
PyMuPDF
python-doc